Rename `credentials` package to `auth` to reflect more general nature
