+++ auth/credentials_test.go
@@ -1,4 +1,4 @@
-package credentials
+package auth
 
 import (
 	"encoding/base64"
